+++ src/main/java/webserver/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
