+++ src/main/java/webserver/response/ResponseHeader.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import org.springframework.http.HttpStatus;
